Allow multiple url parameter
Imagine you want add more than one parameter (q) to the uri, you need to manually construct
the whole ajax request. This can be simplified by checking the given url for a question mark
`example.com?address=Foo`. In this case append an ambersand (&) to the end of the url to construct this path
`example.com?address=Foo&q=Bar`.